pyproject.toml
[project]
name = "clashking-bot"
version = "1.0.0"
description = "ClashKing Discord Bot"
requires-python = ">=3.13"
dependencies = [
    "aiocache",
    "apscheduler",
    "ballpark @ git+https://github.com/Mee6/python-ballpark.git",
    "chat-exporter",
    "clashy-py",
    "coc.py",
    "aiohttp",
    "disnake",
    "disnake-ext-fluent @ git+https://github.com/MagicTheDev/disnake-ext-fluent.git",
    "dnspython",
    "emoji",
    "expiring-dict",
    "fastapi",
    "fluent-runtime",
    "hashids",
    "imagehash",
    "ipython",
    "kaleido",
    "loguru",
    "matplotlib",
    "meilisearch-python-sdk",
    "motor",
    "msgspec",
    "numerize",
    "numpy",
    "openai",
    "openpyxl",
    "orjson",
    "pandas",
    "pendulum",
    "plotly",
    "pymitter",
    "pymongo",
    "python-dateutil",
    "python-dotenv",
    "python-snappy",
    "pytz",
    "redis",
    "requests>=2.32.5",
    "sentry-sdk",
    "stringcase",
    "ujson",
    "uvicorn",
    "wand",
    "websockets",
    "xlsxwriter",
]

[tool.blue]
line-length = 150

[tool.isort]
profile = "black"
line_length = 150
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
lines_after_imports = 2
